refactoring v3 module
