+++ apisix/admin/services.lua
@@ -19,6 +19,7 @@ local get_routes = require("apisix.router").http_routes
 local apisix_upstream = require("apisix.upstream")
 local schema_plugin = require("apisix.admin.plugins").check_schema
 local utils = require("apisix.admin.utils")
+local v3_adapter = require("apisix.admin.v3_adapter")
 local tostring = tostring
 local ipairs = ipairs
 local type = type
@@ -146,7 +147,7 @@ function _M.get(id)
     end
 
     utils.fix_count(res.body, id)
-    utils.pagination(res.body)
+    v3_adapter.pagination(res.body)
     return res.status, res.body
 end
