+++ apisix/admin/stream_routes.lua
@@ -17,6 +17,7 @@
 local core = require("apisix.core")
 local utils = require("apisix.admin.utils")
 local stream_route_checker = require("apisix.stream.router.ip_port").stream_route_checker
+local v3_adapter = require("apisix.admin.v3_adapter")
 local tostring = tostring
 
 
@@ -114,7 +115,7 @@ function _M.get(id)
     end
 
     utils.fix_count(res.body, id)
-    utils.pagination(res.body)
+    v3_adapter.pagination(res.body)
     return res.status, res.body
 end
