+++ apisix/core/etcd.lua
@@ -21,8 +21,7 @@
 
 local fetch_local_conf  = require("apisix.core.config_local").local_conf
 local array_mt          = require("apisix.core.json").array_mt
-local try_read_attr     = require("apisix.core.table").try_read_attr
-local log               = require("apisix.core.log")
+local v3_adapter        = require("apisix.admin.v3_adapter")
 local etcd              = require("resty.etcd")
 local clone_tab         = require("table.clone")
 local health_check      = require("resty.etcd.health_check")
@@ -35,55 +34,6 @@ local ngx_config_prefix = ngx.config.prefix()
 
 local is_http = ngx.config.subsystem == "http"
 local _M = {}
-
-
-local admin_api_version
-local function is_v3()
-    if admin_api_version then
-        if admin_api_version == "v3" then
-            return true
-        end
-
-        if admin_api_version == "default" then
-            return false
-        end
-    end
-
-    local local_conf, err = fetch_local_conf()
-    if not local_conf then
-        admin_api_version = "default"
-        log.error("failed to fetch local conf: ", err)
-        return false
-    end
-
-    local api_ver = try_read_attr(local_conf, "apisix", "admin_api_version")
-    if api_ver ~= "v3" then
-        admin_api_version = "default"
-        return false
-    end
-
-    admin_api_version = api_ver
-    return true
-end
-
-
-local function to_v3(body, action)
-    if not is_v3() then
-        body.action = action
-    end
-end
-
-
-local function to_v3_list(body)
-    if not is_v3() then
-        return
-    end
-
-    if body.node.dir then
-        body.list = body.node.nodes
-        body.node = nil
-    end
-end
 
 
 -- this function create the etcd client instance used in the Admin API
@@ -219,7 +169,7 @@ function _M.get_format(res, real_key, is_dir, formatter)
         return not_found(res)
     end
 
-    to_v3(res.body, "get")
+    v3_adapter.to_v3(res.body, "get")
 
     if formatter then
         return formatter(res)
@@ -247,7 +197,7 @@ function _M.get_format(res, real_key, is_dir, formatter)
     end
 
     res.body.kvs = nil
-    to_v3_list(res.body)
+    v3_adapter.to_v3_list(res.body)
     return res
 end
 
@@ -324,7 +274,7 @@ local function set(key, value, ttl)
     res.headers["X-Etcd-Index"] = res.body.header.revision
 
     -- etcd v3 set would not return kv info
-    to_v3(res.body, "set")
+    v3_adapter.to_v3(res.body, "set")
     res.body.node = {}
     res.body.node.key = prefix .. key
     res.body.node.value = value
@@ -387,7 +337,7 @@ function _M.atomic_set(key, value, ttl, mod_revision)
 
     res.headers["X-Etcd-Index"] = res.body.header.revision
     -- etcd v3 set would not return kv info
-    to_v3(res.body, "compareAndSwap")
+    v3_adapter.to_v3(res.body, "compareAndSwap")
     res.body.node = {
         key = key,
         value = value,
@@ -425,7 +375,7 @@ function _M.push(key, value, ttl)
         return nil, err
     end
 
-    to_v3(res.body, "create")
+    v3_adapter.to_v3(res.body, "create")
     return res, nil
 end
 
@@ -449,7 +399,7 @@ function _M.delete(key)
     end
 
     -- etcd v3 set would not return kv info
-    to_v3(res.body, "delete")
+    v3_adapter.to_v3(res.body, "delete")
     res.body.node = {}
     res.body.key = prefix .. key
